Let MainCommonApi extends BaseApi
