+++ biz/main/src/main/kotlin/io/goooler/demoapp/main/api/BaseApi.kt
@@ -1,0 +1,15 @@
+package io.goooler.demoapp.main.api
+
+import io.goooler.demoapp.base.util.ParamMap
+import io.goooler.demoapp.main.bean.MainRepoListBean
+import retrofit2.http.GET
+import retrofit2.http.Path
+import retrofit2.http.QueryMap
+
+interface BaseApi {
+  @GET("users/{user}/repos")
+  suspend fun getRepoList(
+    @Path("user") user: String,
+    @QueryMap params: ParamMap,
+  ): List<MainRepoListBean>
+}
