+++ biz/main/src/main/kotlin/io/goooler/demoapp/main/api/MainCommonApi.kt
@@ -1,16 +1,3 @@
 package io.goooler.demoapp.main.api
 
-import io.goooler.demoapp.base.util.ParamMap
-import io.goooler.demoapp.main.bean.MainRepoListBean
-import retrofit2.http.GET
-import retrofit2.http.Path
-import retrofit2.http.QueryMap
-
-interface MainCommonApi {
-
-  @GET("users/{user}/repos")
-  suspend fun getRepoList(
-    @Path("user") user: String,
-    @QueryMap params: ParamMap,
-  ): List<MainRepoListBean>
-}
+interface MainCommonApi : BaseApi
